1 file modified, 14 added, 15 deleted: updated classes (v2) & new display functions
