+++ src/app/classes/anthill.ts
