+++ src/app/classes/pheromone.ts
