+++ src/app/classes/supply.ts
