--- src/app/logic/README.md
@@ -1,1 +1,0 @@
-# AntDynamics
